Refactor: make part-finder easier to handle, extract the part-moving editor function and the component-listing logic
